release: prepare 0.1.2
- Bump version to 0.1.2 in CMakeLists.txt and library.json
- Raise cmake_minimum_required to 3.15 to fix silent suppression of
  -Wall/-Wextra/-Wpedantic (C_COMPILER_ID comma form is a no-op on 3.13/3.14)
- Change COMPATIBILITY to SameMinorVersion (SameMajorVersion was too
  permissive for a pre-1.0 library)
- mazarbulib_set_screen: change parameter type uint8_t -> int to match
  the int return type of mazarbulib_register_screen; update doc comment
  to document the NULL-ctx no-op behaviour
- Remove unreachable NULL guard in MAZARBULIB_TYPE_STRING branch
  (value_ptr is already validated in mazarbulib_register_row)
- Fix navigation footer format: '[%c]ext  [%c]rev' -> '[%c]=next  [%c]=prev'
  so labels are correct for any custom MAZARBULIB_NAV_NEXT/PREV binding
- Fix misleading MAZARBULIB_TITLE_MAX_LEN_ comment (overhead is 12 bytes,
  not '11 + 1')
- Fix MAZARBULIB_LINE_BUF_SIZE_ comment (was split mid-string across lines)
- examples/posix.c, examples/stm32_hal.c: mark uart_send and
  terminal_clear static; remove unreachable return after for(;;);
  add assert(len <= UINT16_MAX) before uint16_t cast in stm32_hal.c;
  fix backslash-in-comment warning in posix.c
- tests: assert on fake_uart_send buffer overflow instead of silently
  dropping data; add test_null_guards covering tick/next/prev/feed_char/
  set_screen with NULL context; register test_null_guards in CMakeLists.txt
- README: fix code block language tag ('"C"' -> 'text'); add return-value
  check to Quick Start mazarbulib_init call; fix float format spec (%f -> %.2f)

+++ include/mazarbulib.h
@@ -125,8 +125,9 @@ void mazarbulib_next_screen(mazarbulib_t *ctx);
 // Goes back to the previous screen, wrapping around to the last.
 void mazarbulib_prev_screen(mazarbulib_t *ctx);
 
-// Jumps to the screen at screen_idx. No-op if screen_idx is out of range.
-void mazarbulib_set_screen(mazarbulib_t *ctx, uint8_t screen_idx);
+// Jumps to the screen at screen_idx. No-op if ctx is NULL or screen_idx is
+// out of range.
+void mazarbulib_set_screen(mazarbulib_t *ctx, int screen_idx);
 
 // Feeds a single received byte from the UART RX handler.
 // Handles MAZARBULIB_NAV_NEXT and MAZARBULIB_NAV_PREV; ignores other bytes.
